+++ src/main/kotlin/io/github/dornol/idkit/flake/SnowflakeIdGenerator.kt
@@ -14,7 +14,7 @@ import java.time.Instant
  *
  * @see <a href="https://github.com/twitter/snowflake/tree/snowflake-2010">Twitter Snowflake</a>
  */
-class SnowflakeIdGenerator(
+open class SnowflakeIdGenerator(
     workerId: Int,
     datacenterId: Int,
     epochStart: Instant = Instant.EPOCH,